finalProject/backend/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.1
drf-spectacular==0.27.0
python-decouple==3.8
gunicorn==21.2.0
